+++ models/User.js
@@ -50,4 +50,3 @@ module.exports.getUserById = function (id, callback) {
 module.exports.getAllUsers = function (callback) {
   User.find({}, callback);
 }
-
