+++ public/css/style.css
@@ -1,10 +1,13 @@
 #body {
   background-color: #1b2431;
-  margin-top:50px;
+  margin-top:100px;
   color: #202020;
   font-family: "Montserrat", "Helvetica", "Open Sans", "Arial";
   font-size: 13px;
   min-height:500px;
+}
+#h1{
+  color:white;
 }
 p,
 figure,
@@ -24,7 +27,9 @@ nav ul li{
 	  padding-right: 10px;
     font-family:"Lucida Console", Monaco, monospace ;
 }
-
+.dropdown-menu{
+  overflow:auto;
+}
 .dropdown-menu li{
     font-family:"Comic Sans MS", cursive, sans-serif;
 }
@@ -43,6 +48,7 @@ nav ul li{
   padding-bottom: 40px;
 }
 .side-bar > li > a {
+  color: #9E9E9E;
   width: 220px;
 }
 #sidebar > ul {
@@ -51,10 +57,10 @@ nav ul li{
 }
 .side-bar > li > a:hover, 
 .side-bar >li > a:focus {
-  background-color: #2298F1;
+  background-color: #E9BF1E;
 }
 .card-list {
-  margin-top: 70px;
+
   margin-bottom:40px;
   width: 100%;
 }
@@ -229,4 +235,37 @@ nav ul li{
 }
 .danger-text {
   color: #A84D43 !important;
+}footer {
+	padding:20px;
+	text-align: center;
+	color:#DCE35B;
+	font-family: tahoma;
+	font-size: 1.3em;
+}
+
+footer a {
+	color: #247BE6;
+}
+
+footer a:hover {
+	color: #74C44F;
+}
+ul.social-buttons li a {
+	display: block;
+	height: 40px;
+	width: 40px;
+	font-size: 20px;
+	line-height: 40px;
+	-webkit-transition: all 0.3s;
+	-moz-transition: all 0.3s;
+	transition: all 0.3s;
+}
+
+ul.social-buttons li a:hover,
+ul.social-buttons li a:focus,
+ul.social-buttons li a:active {
+	background-color:#34495e;
+}
+.love{
+  color:#FF2B56;
 }
