+++ views/partials/footer.ejs
@@ -1,8 +1,15 @@
-<!--
-    script tags
--->
-<script src='js/jquery.min.js'></script>
-<script src='js/bootstrap.min.js'></script>
+<footer class="text-center">
 
+    <div class=''>
+        <a href="https://twitter.com/sparkplugnow" target='blank'>
+                        <i class="fa fa-twitter"></i>
+                    </a> | Crafted and Designed with
+        <i class='fa fa-heart love'></i> by
+        <a href="https://sparkplugnow.com">
+                            Sparkplugnow
+                        </a>
+    </div>
+</footer>
 </body>
+
 </html>
