--- views/transaction.ejs
@@ -1,11 +1,0 @@
-<%include partials/header.ejs%>
-    <div class='container'>
-    <% transactions.forEach(function(transaction){%>
-        <div>
-<% transaction.from%> sent
-    <% transaction.amount%> coins to <% transaction.to%>
-        </div>
-         
-    <% })%>
-    </div>
-<%include partials/footer.ejs%>
